web/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
Click==7.0
Django==2.2.7
openslide-python==1.1.1
pathspec==0.6.0
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
regex==2019.11.1
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
uWSGI==2.0.18
